+++ CHANGES.rst
@@ -3,4 +3,10 @@ Changeslog
 
 Version 0.2.2
 -------------
-* Define the order_block, order_inline and order_initial attributes inside the __init__ method in the Lexer class.
+* Define the *order_block*, *order_inline* and *order_initial* attributes inside the __init__ method in the Lexer class.
+
+Version 0.2.3
+-------------
+* Redefine the *order_block*, *order_inline* and *order_initial* attributes inside the __init__ method in the Lexer class,
+  because no run correctly. Now you can use *order_block* and *order_inline* to add the new grammar.
+* Change in the *header2* grammar element.